♻️ refactor: likeCount 최신값 보장

+++ src/main/java/troublog/backend/domain/trouble/repository/LikeRepository.java
@@ -15,5 +15,5 @@ public interface LikeRepository extends JpaRepository<Like, Long> {
 
 	Optional<Like> findByUserIdAndPostId(Long userId, Long postId);
 
-	long countByPostId(Long postId);
+	int countByPostId(Long postId);
 }
